requirements.txt
python-telegram-bot>=22.0
fastapi>=0.95
uvicorn>=0.23.2
gunicorn
